Login and register rework
Added custom hook to refresh token ( need to discuss authentication logic)

seperated login and register

added simple cookies and logout (need to change it in future)
